// src/app/shared/enums/QuestionType.ts
export enum QuestionType {
	TPO_READING_SINGLE_TYPE = "TPO_READING_SINGLE_CHOICE_TYPE",
	TPO_READING_INSERT_TYPE = "TPO_READING_INSERT_CHOICE_TYPE",
	TPO_READING_TOPIC_TYPE = "TPO_READING_TOPIC_TYPE",
	TPO_READING_MULTIPLE_TYPE = "TPO_READING_MULTIPLE_TYPE",
	TPO_READING_CATEGORY_TYPE = "TPO_READING_CATEGORY_TYPE",
	VOCABULARY_TYPE = "VOCABULARY_TYPE",
	INTEGRATED_WRITING_TYPE = "INTEGRATED_WRITING_TYPE",
	INDEPENDENT_WRITING_TYPE = "INDEPENDENT_WRITING_TYPE",
	TPO_LISTENING_SINGLE_CHOICE_TYPE = "TPO_LISTENING_SINGLE_CHOICE_TYPE",
	TPO_LISTENING_MULTIPLE_CHOICE_TYPE = "TPO_LISTENING_MULTIPLE_CHOICE_TYPE",
	TPO_LISTENING_TABLE_CHOICE_TYPE = "TPO_LISTENING_TABLE_CHOICE_QUESTION",
	TPO_LISTENING_REPEAT_TYPE = "TPO_LISTENING_REPEAT_QUESTION",
	TPO_LISTENING_SEQUENCE_TYPE = "TPO_LISTENING_SEQUENCE_TYPE",
	TPO_SPEAKING_Q1Q2_TYPE = "TPO_SPEAKING_Q1Q2_QUESTION",
	TPO_SPEAKING_Q3Q4_TYPE = "TPO_SPEAKING_Q3Q4_QUESTION",
	TPO_SPEAKING_Q5Q6_TYPE = "TPO_SPEAKING_Q5Q6_QUESTION"

}